Add support for logging to a secondary non fluent log
There are various issues with using ActiveSupport::Logger.broadcast in
conjunction with act-fluent-logger-rails to log to more than one log.

See: https://github.com/actindi/act-fluent-logger-rails/issues/19

This allows you to add an optional parameter that it is a new logfile.

One caveat is that it will log at the end of the request along with
fluentd, and be in the same format as your fluentd log. (Which you may
    or may not want.)